requirements.txt
numpy==1.23.5
paho-mqtt==1.6.1
